% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burd_anninc.R
\name{burd_anninc}
\alias{burd_anninc}
\title{Burdened Income
Minimum annual income required for total housing costs to remain at or below 50\% of household income:}
\usage{
burd_anninc(monthlypayment)
}
\arguments{
\item{monthlypayment}{Total Monthly payments towards home-ownership}
}
\value{
Minimum Annual Income required to keep \code{monthlypayment} at or below 50\% of income
}
\description{
Burdened Income
Minimum annual income required for total housing costs to remain at or below 50\% of household income:
}
\examples{

burd_anninc(2000)
}
