% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burd_income.R
\name{burd_income}
\alias{burd_income}
\title{Burdened Income}
\usage{
burd_income(monthlypayment)
}
\arguments{
\item{monthlypayment}{Total Monthly payments towards home-ownership}
}
\value{
Minimum Annual Income required to keep \code{monthlypayment} at or below 50\% of income
}
\description{
Minimum annual income required for total housing costs to remain at or below 50\% of household income:
}
\details{
\itemize{
\item To be deprecated. Use reqd_income() instead *
}
}
\examples{

burd_income(2000)
}
