% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downpayment.R
\name{downpayment}
\alias{downpayment}
\title{Down-payment Amount}
\usage{
downpayment(home_price, mortgage_amount = 0.9)
}
\arguments{
\item{home_price}{Sale Price of property in question}

\item{mortgage_amount}{Mortgage as a proportion (out of 1) of sale price (default: 0.90)}
}
\value{
The down-payment based on what proportion (\code{mortgage_amount}) of the \code{home_price} will be paid
}
\description{
Total downpayment based on sale price and mortgage amount (out of 1)
}
\examples{
downpayment(home_price = 250000,
            mortgage_amount = 0.90)
}
