% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downpayment_amt.R
\name{downpayment_amt}
\alias{downpayment_amt}
\title{Down-payment Amount}
\usage{
downpayment_amt(home_price, downpayment_rate = 0.1)
}
\arguments{
\item{home_price}{Sale Price of property in question}

\item{downpayment_rate}{Downpayment as a proportion (out of 1) of sale price (default: 0.10)}
}
\value{
The down-payment based on what proportion (\code{mortgage_amount}) of the \code{home_price} will be paid
}
\description{
Total downpayment based on sale price and mortgage amount (out of 1)
}
\examples{
downpayment_amt(home_price = 250000,
            downpayment_rate = 0.10)
}
