% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_homeprice.R
\name{max_homeprice}
\alias{max_homeprice}
\title{Affordable Home Price based on AMI - Area Median Income}
\usage{
max_homeprice(
  mortgage_rate,
  property_tax,
  annualinc = 69037.5,
  max_monthly_rate = 0.3,
  homeowners_ins = 1000,
  condo_fee = 1000,
  downpayment_rate = 0.1,
  loan_term = 30,
  pmi_factor = 7.5
)
}
\arguments{
\item{mortgage_rate}{Fixed monthly mortgage interest rate in percentage (out of 100)}

\item{property_tax}{Property tax rate per thousand (out of 1000) for the given area -
NOTE: Generally property taxes are always mentioned per thousand}

\item{annualinc}{Household Income (default is median for Boston Metro 2019 ACS 5yr estimate)}

\item{max_monthly_rate}{Maximum monthly rate (out of 1) towards homeownership costs (default: 0.3)}

\item{homeowners_ins}{Annual home owners insurance costs (default: 1000)}

\item{condo_fee}{Annual condo fee (default: 1000)}

\item{downpayment_rate}{Downpayment as a proportion (out of 1) of sale price (default: 0.10)}

\item{loan_term}{Duration of loan payback time in years (default: 30)}

\item{pmi_factor}{For loans with less than 20\% downpayment, mortgage insurance rate in percent (out of 100) (default: 7.5)}
}
\description{
AMI for the Boston-Cambridge-Newton MA-NH Metro Area 2019 ACS-5 Year Estimates at: 86274 USD
https://data.census.gov/cedsci/table?q=area\%20median\%20income&g=314M500US1446014454,1446015764&tid=ACSST5Y2019.S1903&hidePreview=true
}
\examples{

max_homeprice(annualinc = c(61000, 75000, 22330, 120000),
              mortgage_rate = 2.7,
              property_tax = 6)
}
