% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_income.R
\name{min_income}
\alias{min_income}
\title{Minimum Non-Burdened Income}
\usage{
min_income(monthlypayment)
}
\arguments{
\item{monthlypayment}{Total Monthly payments towards home-ownership}
}
\value{
Minimum Annual Income required to keep \code{monthlypayment} at or below 30\% of income
}
\description{
Minimum annual income required for total housing costs to remain at or below 30\% of household income:
}
\examples{

min_income(2000)
}
