% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlypayment.R
\name{monthlypayment}
\alias{monthlypayment}
\title{Monthly Payments}
\usage{
monthlypayment(
  mortgage_rate,
  property_tax,
  home_price,
  homeowners_ins = 1000,
  condo_fee = 1000,
  mortgage_amount = 0.9,
  loan_term = 30,
  pmi_factor = 7.5
)
}
\arguments{
\item{mortgage_rate}{Fixed monthly mortgage interest rate in percentage (out of 100)}

\item{property_tax}{Property tax rate per thousand (out of 1000) for the given area -
NOTE: Generally property taxes are always mentioned per thousand}

\item{home_price}{Sale Price of property in question}

\item{homeowners_ins}{Annual home owners insurance costs (default: 1000)}

\item{condo_fee}{Annual condo fee (default: 1000)}

\item{mortgage_amount}{Mortgage as a proportion (out of 1) of sale price (default: 0.90)}

\item{loan_term}{Duration of loan payback time in years (default: 30)}

\item{pmi_factor}{For loans with less than 20\% downpayment, mortgage insurance rate in percent (out of 100) (default: 7.5)}
}
\description{
Function to calculate monthly home-ownership costs based on tax rates, interest rates and maintenance rates
}
\examples{
monthlypayment(mortgage_rate = 2.7,
               property_tax = 6,
               home_price = 350000)
}
