% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqd_income.R
\name{reqd_income}
\alias{reqd_income}
\title{Minimum Required Income for Homeownership}
\usage{
reqd_income(monthlypayment, housing_cost_rate = 0.3)
}
\arguments{
\item{monthlypayment}{Total Monthly payments towards home-ownership}

\item{housing_cost_rate}{Monthly rate (out of 1) towards home-ownership costs (default: 0.3)}
}
\value{
Minimum Annual Income required to keep \code{monthlypayment} at or below set percent threshold of household income:
}
\description{
Minimum annual income required for total housing costs to remain at or below set percent threshold of household income:
}
\examples{
reqd_income(monthlypayment = 2000,
            housing_cost_rate = 0.30)
}
